Basic Structure of Recipes Website
The first project on The Odin Project was to build a simple recipes
website using plain HTML to learn how a website is structured.
I used anchor links to connect different recipes with the Index.html
and unordered lists as well as ordered lists to list the ingredients
and steps.